+++ jetforce/server.py
@@ -145,6 +145,8 @@ class GeminiServer(Factory):
         """
         self.log_message(ABOUT)
         self.log_message(f"Server hostname is {self.hostname}")
+        if hasattr(self.app, "root"):
+            self.log_message(f"Root directory: {self.app.root}")
         if self.proxy_protocol:
             self.log_message("PROXY protocol is enabled")
         if self.use_tls: